Move SettingStore to useSettings
