--- app/src/stores/SettingStore.ts
@@ -1,19 +1,0 @@
-import { makeObservable, observable } from "mobx"
-import { makePersistable } from "mobx-persist-store"
-import { Language } from "../localize/useLocalization"
-
-export default class SettingStore {
-  language: Language | null = null
-
-  constructor() {
-    makeObservable(this, {
-      language: observable,
-    })
-
-    makePersistable(this, {
-      name: "SettingStore",
-      properties: ["language"],
-      storage: window.localStorage,
-    })
-  }
-}
